requirements.txt
# FastAPI and web service dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Core dependencies for Gomoku AI Agent
numpy>=1.21.0
python-dotenv>=0.19.0

# Gomoku framework (install from local gomoku-ai directory or repository)
# Note: The gomoku package appears to be a local dependency
# Install with: pip install -e ../gomoku-ai/ (adjust path as needed)

# HTTP client for testing and service communication
httpx>=0.25.0
aiohttp>=3.8.0

# Optional: If using specific OpenAI client or other LLM providers
openai>=1.0.0
requests>=2.28.0

# Development and testing dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
